reducers: extract pageReducer to its own file
